refactor: move SSRF validation to utils/url.ts and expand blocklist
Address PR review feedback:
- Move isAllowedHealthCheckUrl from config/url-validation.ts to utils/url.ts,
  colocated with existing normalizeOptionalHttpUrl
- Reuse normalizeOptionalHttpUrl for protocol validation instead of duplicating
- Add GCP metadata (metadata.google.internal) and AWS IPv6 metadata
  (fd00:ec2::254) to blocklist
- Add test for private network URLs (192.168.x, 10.x) as allowed targets
